Rearrange more support code to be in the support module
