+++ spinnaker_proxy/support.py
@@ -13,11 +13,60 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-""" Simple support classes.
+""" Simple support classes and function.
 """
 
-import struct
 from abc import abstractmethod, ABCMeta as Abstract
+import socket
+import struct
+
+
+def udp_socket(bind_port=None, connect_address=None):
+    """ How to make a UDP socket.
+
+    :param int bind_port:
+        If provided, what local port to send and receive packets via.
+    :param tuple(str,int) connect_address:
+        If provided, what remote IP address/port to send packets to and
+        receive them from.
+    :return: The configured socket.
+    :rtype: socket.SocketType
+    """
+    sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    try:
+        if bind_port is not None:
+            sock.bind(("", bind_port))
+        if connect_address is not None:
+            sock.connect(connect_address)
+    except Exception as e:
+        sock.close()
+        raise e
+    return sock
+
+
+def tcp_socket(bind_port=None, connect_address=None):
+    """ How to make a TCP socket.
+
+    :param int bind_port:
+        If provided, what local port to send and receive data via.
+    :param tuple(str,int) connect_address:
+        If provided, what remote IP address/port to send data to and
+        receive it from.
+    :return: The configured socket.
+    :rtype: socket.SocketType
+    """
+    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    try:
+        if bind_port is not None:
+            sock.setsockopt(
+                socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+            sock.bind(("", bind_port))
+        if connect_address is not None:
+            sock.connect(connect_address)
+    except Exception as e:
+        sock.close()
+        raise e
+    return sock
 
 
 class TCPDatagramProtocol(object):
@@ -81,7 +130,7 @@ class DatagramProxy(object, metaclass=Abstract):
         :return:
             All sockets should be selected for readability, and, when
             readable, that callback should be called.
-        :rtype: dict(int, ~collections.abc.Callable)
+        :rtype: dict(socket.SocketType, ~collections.abc.Callable)
         """
         raise NotImplementedError
